Uppdaterat aktivitet

+++ src/BUPSystem/ActivityGUI/Activity register.xaml.cs
@@ -11,10 +11,8 @@ namespace BUPSystem.ActivityGUI
     {
         public ObservableCollection<Activity> ActivityList
         {
-            get
-            {
-                return new ObservableCollection<Activity>(ActivityManagement.Instance.GetActivities());
-            }
+            get { return ActivityManagement.Instance.Activities; }
+            set { ActivityManagement.Instance.Activities = value; }
         }
 
         /// <summary>
@@ -23,7 +21,6 @@ namespace BUPSystem.ActivityGUI
         public ActivityRegister()
         {
             InitializeComponent();
-
             DataContext = this;
         }
 
@@ -36,6 +33,11 @@ namespace BUPSystem.ActivityGUI
         {
             ActivityManager am = new ActivityManager();
             am.ShowDialog();
+            if (am.DialogResult == true)
+            {
+                ActivityManagement.Instance.AddActicity(am.Activity);
+                //lblInfo.Content = "Ny användare skapad";
+            }
         }
 
         /// <summary>
@@ -45,8 +47,20 @@ namespace BUPSystem.ActivityGUI
         /// <param name="e"></param>
         private void btnChange_Click(object sender, RoutedEventArgs e)
         {
-            ActivityManager am = new ActivityManager(ActivityList[lvActivityRegister.SelectedIndex]);
-            am.ShowDialog();
+            // Make sure the sure the user has selected an item in the listview
+            if (lvActivityRegister.SelectedItem != null)
+            {
+                ActivityManager am = new ActivityManager(ActivityList[lvActivityRegister.SelectedIndex]);
+                am.ShowDialog();
+
+                if (am.DialogResult == true)
+                {
+                    ActivityManagement.Instance.Update();
+                    //lblInfo.Content = "Ny användare skapad";
+                }
+            }
+            else
+                MessageBox.Show("Markera en aktivitet i listan", "Ingen vald aktivitet");     
         }
 
         /// <summary>
@@ -56,7 +70,21 @@ namespace BUPSystem.ActivityGUI
         /// <param name="e"></param>
         private void btnDelete_Click(object sender, RoutedEventArgs e)
         {
-            ActivityManagement.Instance.DeleteActivity(ActivityList[lvActivityRegister.SelectedIndex]);
+            // Make sure the sure the user has selected an item in the listview
+            if (lvActivityRegister.SelectedItem != null)
+            {
+                // Confirm that the user wishes to delete
+                MessageBoxResult mbr = MessageBox.Show("Vill du verkligen ta bort den här aktiviteten?",
+                    "Ta bort aktivitet", MessageBoxButton.YesNo);
+
+                if (mbr == MessageBoxResult.Yes)
+                {
+                    ActivityManagement.Instance.DeleteActivity(ActivityList[lvActivityRegister.SelectedIndex]);
+                    //lblInfo.Content = "Användaren togs bort";
+                }
+            }
+            else
+                MessageBox.Show("Markera en aktivitet att ta bort", "Ingen vald aktivitet");
         }
     }
 }
